Since the subprocess.Popen.poll() call checks if the child process has terminated, we can use a flag to make sure to not terminate the job twice with the job.terminate() call after `if has_error`.

+++ custodian/custodian.py
@@ -458,6 +458,7 @@ class Custodian:
             # corrections.
             has_error = False
             zero_return_code = True
+            job_alive = True
 
             # While the job is running, we use the handlers that are
             # monitors to monitor the job.
@@ -468,6 +469,7 @@ class Custodian:
                         n += 1
                         time.sleep(self.polling_time_step)
                         if p.poll() is not None:
+                            job_alive = False
                             break
                         terminate = self.terminate_func or p.terminate
                         if n % self.monitor_freq == 0:
@@ -491,7 +493,7 @@ class Custodian:
             else:
                 has_error = self._do_check(self.handlers)
 
-            if has_error:
+            if has_error and job_alive:
                 # This makes sure the job is killed cleanly for certain systems.
                 job.terminate()
